pyproject.toml
[tool.poetry]
name = "pp_data"
version = "0.1.0"
description = ""
authors = ["Dmitriy Grigorev"]

[tool.poetry.dependencies]
python = "^3.8"
dvc = "^2.10.0"
hydra-core = "^1.1.1"
pdoc3 = "^0.10.0"
install = "^1.3.5"
vaex-core = "^4.9.1"
vaex-viz = "^0.5.1"
vaex-server = "^0.8.1"
vaex-hdf5 = "^0.12.1"
vaex-ml = "^0.17.0"
python-dotenv = "^0.20.0"
requests = "^2.27.1"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build   
)/ 
'''
